Move whitespace
