+++ lib/src/parser/literals.rs
@@ -1,6 +1,4 @@
 //! Tokens and literals
-#[macro_use]
-pub mod whitespace;
 
 pub mod key;
 pub mod number;
@@ -9,5 +7,3 @@ pub mod number;
 pub use key::{key, Key};
 #[doc(inline)]
 pub use number::{number, Number};
-#[doc(inline)]
-pub use whitespace::{inline_whitespace, newline, whitespace};
