+++ lib/src/parser/object.rs
@@ -17,7 +17,7 @@ use nom::{alt, call, char, do_parse, named, peek, recognize, tag, terminated, IR
 
 use crate::parser::expression::{expression, Expression};
 use crate::parser::identifier::identifier;
-use crate::parser::literals::newline;
+use crate::parser::whitespace::newline;
 use crate::HashMap;
 
 // TODO: Dealing with expressions and ambiguity. See reference
